create order.service.ts to handle order requests
